Move UserDefinedLogicalNode to extension.rs
